Cart updates

+++ unipantry/src/app/cart/cart.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit, AfterViewInit, OnChanges } from '@angular/core';
+import { Component, AfterViewInit, OnInit } from '@angular/core';
 import { CartService } from '../cart.service';
 import { Product } from '../product';
 
@@ -7,7 +7,7 @@ import { Product } from '../product';
   templateUrl: './cart.component.html',
   styleUrls: ['./cart.component.css']
 })
-export class CartComponent implements OnInit, AfterViewInit, OnChanges {
+export class CartComponent implements AfterViewInit, OnInit {
 
   cart: Product[] = [];
 
@@ -15,25 +15,28 @@ export class CartComponent implements OnInit, AfterViewInit, OnChanges {
 
   cartQuantity = 0;
 
-  total: number;
+  total = 0;
 
   giftCredit: number;
 
-  subtotal: number;
+  giftCreditUsed = false;
+
+  subtotal = 0;
+
+  totalRemaining = 30;
 
   constructor(private cartService: CartService) { }
 
   ngOnInit() {
-    this.cart = [];
     this.getCart();
   }
 
   ngAfterViewInit() {
-    this.setQuantities(-1);
-  }
-
-  ngOnChanges() {
-    this.getCart();
+    this.setQuantities(-1, this.cart, this.productDict);
+    let i: number;
+    for (i = 0; i < this.cart.length; i++) {
+      this.setQuantities(i, this.cart, this.productDict);
+    }
   }
 
   getCart() {
@@ -46,32 +49,39 @@ export class CartComponent implements OnInit, AfterViewInit, OnChanges {
         this.giftCredit = cart['credit'];
         this.subtotal = cart['subtotal'];
       });
+    if (this.giftCredit > this.total) {
+      this.giftCredit = this.total;
+    }
+    this.totalRemaining = 30 - this.total;
   }
 
   removeFromCart(product: Product, quantity: number) {
+    if (quantity >= this.productDict[product.name]) {
+      this.removeAll(product);
+      return;
+    }
     this.cartService.removeFromCart(product, quantity);
     this.getCart();
     const index = this.cart.indexOf(product);
     if (index !== -1) {
-      this.setQuantities(index);
+      this.setQuantities(index, this.cart, this.productDict);
     }
   }
 
   removeAll(product: Product) {
     this.cartService.removeAll(product);
     this.getCart();
+    this.setQuantities(-1, this.cart, this.productDict);
+  }
+
+  useGiftCredit() {
+    this.giftCreditUsed = true;
   }
 
   addToCart(product: Product, quantity: number) {
     this.cartService.addToCart(product, quantity);
-    console.log(this.cart);
-    if (!this.cart.includes(product)) {
-      this.getCart();
-      this.setQuantities(-1);
-    } else {
-      this.getCart();
-      this.setQuantities(this.cart.indexOf(product));
-    }
+    this.getCart();
+    this.setQuantities(this.cart.indexOf(product), this.cart, this.productDict);
   }
 
   updateQuantity(product: Product) {
@@ -87,11 +97,16 @@ export class CartComponent implements OnInit, AfterViewInit, OnChanges {
     }
   }
 
-  setQuantities(index: number) {
+  setQuantities(index: number, cart: Product[], productDict: {}) {
     (document.getElementById('cart-quantity') as any).textContent = this.cartQuantity;
     if (index !== -1) {
-      const query = document.querySelectorAll('.quantity');
-      (query[index] as any).value = this.productDict[this.cart[index].name];
+      const checkExist = setInterval(function () {
+        if (document.querySelectorAll('.quantity').length === cart.length) {
+          const query = document.querySelectorAll('.quantity');
+          (query[index] as any).value = productDict[cart[index].name];
+          clearInterval(checkExist);
+        }
+      }, 100);
     }
   }
 }
